backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
PyJWT==2.8.0

# Oracle Database Support
cx-Oracle==8.3.0
oracledb==1.4.0

# Data Processing and Export
pandas==2.1.3
openpyxl==3.11.0
python-dateutil==2.8.2

# Report Generation
reportlab==4.0.7
Jinja2==3.1.2

# Async Task Processing (optional: Celery or RQ)
celery==5.3.4
redis==5.0.1

# Security and Validation
email-validator==2.1.0
bcrypt==4.1.2
slowapi==0.1.9

# Testing
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
anyio==3.7.1

# Monitoring and Logging
python-logging-loki==0.3.2

# API Documentation

